# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TrustFirst freelance marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/{jobId}/proposals:
    get:
      operationId: getJobProposals
      tags: [proposals]
      summary: Get proposals for a job
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of proposals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Proposal"
  /proposals:
    post:
      operationId: createProposal
      tags: [proposals]
      summary: Submit a proposal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalInput"
      responses:
        "201":
          description: Proposal submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
  /proposals/mine:
    get:
      operationId: getMyProposals
      tags: [proposals]
      summary: Get my submitted proposals
      security:
        - bearerAuth: []
      responses:
        "200":
          description: My proposals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Proposal"
  /proposals/{id}/status:
    patch:
      operationId: updateProposalStatus
      tags: [proposals]
      summary: Update proposal status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
...
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        budgetType:
          type: string
          enum: [fixed, hourly]
        budget:
          type: ["number", "null"]
        duration:
          type: ["string", "null"]
        experienceLevel:
          type: ["string", "null"]
        skills:
          type: ["string", "null"]
        status:
          type: string
        clientId:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, title, description, category, budgetType, clientId, createdAt]
...
    Proposal:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        freelancerId:
          type: integer
        coverLetter:
          type: string
        bidAmount:
          type: ["number", "null"]
        deliveryDays:
          type: integer
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, jobId, freelancerId, coverLetter, deliveryDays, status, createdAt]
    ProposalInput:
      type: object
      required: [jobId, coverLetter, bidAmount, deliveryDays]
      properties:
        jobId:
          type: integer
        coverLetter:
          type: string
          minLength: 20
        bidAmount:
          type: number
        deliveryDays:
          type: integer
    ProposalStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted, rejected, withdrawn]
    Project:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: ["integer", "null"]
        clientId:
          type: integer
        freelancerId:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
        budget:
          type: number
        deadline:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        progress:
          type: ["integer", "null"]
      required: [id, clientId, freelancerId, title, description, status, budget, createdAt]
...